Extracted ZoomableSVG feature to separate component. Added ability to pass props to Map class.
