+++ client/src/components/Map.tsx
@@ -1,42 +1,20 @@
 import React, { Component, useRef, useState, useEffect } from 'react';
 import * as d3 from 'd3';
+import ZoomableSVG from './ZoomableSVG';
 
+type Prop = {
+    layoutImage: string
+}
 
-export default class Map extends Component{
+export default class Map extends Component<Prop>{
 
     mapRef!: SVGSVGElement
-
-    private ZoomableSVG( { children, width, height } ){
-
-        const svgRef = useRef()
-        const [k, setK] = useState(1)
-        const [x, setX] = useState(0)
-        const [y, setY] = useState(0)
-
-        useEffect(() => {
-            const zoom = d3.zoom().on("zoom", (event) => {
-                const { x, y, k } = event.transform
-                setK(k)
-                setX(x)
-                setY(y)
-            })
-
-            d3.select(svgRef.current).call(zoom)
-        }, [])
-
-
-        return (
-            <svg ref={svgRef} width={width} height={height}>
-                <g transform={`translate(${x},${y})scale(${k})`}>{children}</g>
-            </svg>
-        )
-    }
 
     private drawMap(){
 
         const svgElement = d3.select(this.mapRef)
         .append("image")
-        .attr("xlink:href", "/submaps/main_floor_0.svg")
+        .attr("xlink:href", `/submaps/${this.props.layoutImage}.svg`)
         .attr("width", 450)
         .attr("height", 300)  
 
@@ -54,7 +32,7 @@ export default class Map extends Component{
             .attr("cx", d3.pointer(event)[0])
             .attr("cy", d3.pointer(event)[1])
             .attr("r", 1.5)
-            .attr("fill","blue")
+            .attr("fill","#41C7F7")
             .attr("stroke-opacity", 1)
             .attr("stroke-width", 1.5)
 
@@ -69,17 +47,19 @@ export default class Map extends Component{
         .attr("cx", x1)
         .attr("cy", y1)
         .attr("r", 1.5)
-        .attr("fill","blue")
+        .attr("fill","#41C7F7")
         .attr("stroke-opacity", 1)
         .attr("stroke-width", 1.5)
+        .classed("animate-pulse", true)
 
         svg.append("circle")
         .attr("cx", x2)
         .attr("cy", y2)
         .attr("r", 1.5)
-        .attr("fill","blue")
+        .attr("fill","#41C7F7")
         .attr("stroke-opacity", 1)
         .attr("stroke-width", 1.5)
+        .classed("animate-pulse", true)
 
         svg.append("line")
         .attr("x1", x1)
@@ -102,16 +82,16 @@ export default class Map extends Component{
     }
 
     render() {
-
+        
         return(
             <>
                 <div className="overflow-hidden whitespace-no-wrap h-full w-full border">
                     <div className="inline-block">
-                        <this.ZoomableSVG width={400} height={300}>
+                        <ZoomableSVG width={400} height={300}>
                             <svg ref={(mapRef: SVGSVGElement) => this.mapRef = mapRef} width={700} height={300}>
                                 
                             </svg>
-                        </this.ZoomableSVG>
+                        </ZoomableSVG>
                     </div>
                 </div>
             </>
